+++ src/lib/Embed.hs
@@ -30,7 +30,8 @@ module Embed (emit, emitTo, emitAnn, emitOp, buildDepEffLam, buildLamAux, buildP
               TraversalDef, traverseDecls, traverseDecl, traverseBlock, traverseExpr,
               clampPositive, buildNAbs, buildNAbsAux, buildNestedLam, zeroAt,
               transformModuleAsBlock, dropSub, appReduceTraversalDef,
-              indexSetSizeE, indexToIntE, intToIndexE, freshVarE) where
+              indexSetSizeE, indexToIntE, intToIndexE, freshVarE,
+              reduceScoped, reduceBlock, reduceAtom, reduceExpr) where
 
 import Control.Applicative
 import Control.Monad
@@ -50,6 +51,7 @@ import Cat
 import Type
 import PPrint
 import Util (bindM2, scanM, restructure)
+import Data.Maybe (fromMaybe)
 
 newtype EmbedT m a = EmbedT (ReaderT EmbedEnvR (CatT EmbedEnvC m) a)
   deriving (Functor, Applicative, Monad, MonadIO, MonadFail, Alternative)
@@ -295,15 +297,19 @@ getSndRef :: MonadEmbed m => Atom -> m Atom
 getSndRef r = emitOp $ SndRef r
 
 -- TODO: refactor?
+-- TODO: is this the best place for the reduction?
 getUnpacked :: MonadEmbed m => Atom -> m [Atom]
-getUnpacked atom = return res where
-  len = case getType atom of
-    TypeCon def params ->
-      let [DataConDef _ bs] = applyDataDefParams def params
-      in length bs
-    RecordTy (NoExt types) -> length types
-    ty -> error $ "Unpacking a type that doesn't support unpacking: " ++ pprint ty
-  res = map (\i -> getProjection [i] atom) [0..(len-1)]
+getUnpacked atom = do
+  scope <- getScope
+  let len = case getType atom of
+        TypeCon def params ->
+          let [DataConDef _ bs] = applyDataDefParams def params
+          in length bs
+        RecordTy (NoExt types) -> length types
+        ty -> error $ "Unpacking a type that doesn't support unpacking: " ++ pprint ty
+      atom' = reduceAtom scope atom
+      res = map (\i -> getProjection [i] atom') [0..(len-1)]
+  return res
 
 app :: MonadEmbed m => Atom -> Atom -> m Atom
 app x i = emit $ App x i
@@ -820,3 +826,61 @@ intToIndexE (VariantTy (NoExt types)) i = do
   start <- Variant (NoExt types) l0 0 <$> intToIndexE ty0 i
   foldM go start zs
 intToIndexE ty _ = error $ "Unexpected type " ++ pprint ty
+
+-- === Reduction ===
+
+reduceScoped :: MonadEmbed m => m Atom -> m (Maybe Atom)
+reduceScoped m = do
+  block <- buildScoped m
+  scope <- getScope
+  return $ reduceBlock scope block
+
+reduceBlock :: Scope -> Block -> Maybe Atom
+reduceBlock scope (Block decls result) = do
+  let localScope = foldMap boundVars decls
+  ans <- reduceExpr (scope <> localScope) result
+  [] <- return $ toList $ localScope `envIntersect` freeVars ans
+  return ans
+
+-- XXX: This should handle all terms of type Type. Otherwise type equality checking
+--      will get broken.
+reduceAtom :: Scope -> Atom -> Atom
+reduceAtom scope x = case x of
+  Var (Name InferenceName _ _ :> _) -> x
+  Var v -> case snd (scope ! v) of
+    -- TODO: worry about effects!
+    LetBound PlainLet expr -> fromMaybe x $ reduceExpr scope expr
+    _ -> x
+  TC con -> TC $ fmap (reduceAtom scope) con
+  Pi (Abs b (arr, ty)) -> Pi $ Abs b (arr, reduceAtom (scope <> (fmap (,PiBound) $ binderAsEnv b)) ty)
+  TypeCon def params -> TypeCon (reduceDataDef def) (fmap rec params)
+  RecordTy (Ext tys ext) -> RecordTy $ Ext (fmap rec tys) ext
+  VariantTy (Ext tys ext) -> VariantTy $ Ext (fmap rec tys) ext
+  ACase _ _ _ -> error "Not implemented"
+  _ -> x
+  where
+    rec = reduceAtom scope
+    reduceNest s n = case n of
+      Empty       -> Empty
+      -- Technically this should use a more concrete type than UnknownBinder, but anything else
+      -- than LetBound is indistinguishable for this reduction anyway.
+      Nest b rest -> Nest b' $ reduceNest (s <> (fmap (,UnknownBinder) $ binderAsEnv b)) rest
+        where b' = fmap (reduceAtom s) b
+    reduceDataDef (DataDef n bs cons) =
+      DataDef n (reduceNest scope bs)
+            (fmap (reduceDataConDef (scope <> (foldMap (fmap (,UnknownBinder) . binderAsEnv) bs))) cons)
+    reduceDataConDef s (DataConDef n bs) = DataConDef n $ reduceNest s bs
+
+reduceExpr :: Scope -> Expr -> Maybe Atom
+reduceExpr scope expr = case expr of
+  Atom val -> return $ reduceAtom scope val
+  App f x -> do
+    let f' = reduceAtom scope f
+    let x' = reduceAtom scope x
+    -- TODO: Worry about variable capture. Should really carry a substitution.
+    case f' of
+      Lam (Abs b (arr, block)) | arr == PureArrow || arr == ImplicitArrow ->
+        reduceBlock scope $ subst (b@>x', scope) block
+      TypeCon con xs -> Just $ TypeCon con $ xs ++ [x']
+      _ -> Nothing
+  _ -> Nothing
